spm-auth-function/requirements.txt
python-jose
pyjwt[crypto]